requirements36-dev.txt
pytest-asyncio
tornado
aiohttp
flake8
-e git://github.com/blazeroot/vcrpy@twisted#egg=vcrpy